# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Aetheris Server API
  version: 1.0.0
  description: Aetheris 服务端设备、事件、访问权限和工作角色 API。
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /api/v1/admin/subjects:
    get:
      operationId: listSubjects
      security: [{ sessionCookie: [] }]
      responses: { '200': { description: 主体列表 } }
...
  /api/v1/admin/effectiveness:
    get:
      operationId: getPersonalEffectiveness
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: subject_id, in: query, required: true, schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        '200': { description: 可解释个人效能报告 }
        '400': { description: 主体或日期范围无效 }
        '403': { description: 无读取权限 }
...
components:
  securitySchemes:
    deviceBearer: { type: http, scheme: bearer }
    sessionCookie: { type: apiKey, in: cookie, name: aetheris_session }
...